backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.13.1  # Database migrations

# Supabase & Auth
supabase==2.3.4
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing

# Validation & Serialization
pydantic[email]==2.5.3
pydantic-settings==2.1.0  # Environment variables

# HTTP Client (for external APIs)
httpx==0.25.2  # Compatible with supabase 2.3.4 (requires <0.26)
aiohttp==3.9.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11

# Monitoring & Logging
python-dotenv==1.0.0
